requirements.txt
appscript==1.2.2
asgiref==3.7.1
certifi==2022.6.15
defusedxml==0.7.1
diff-match-patch==20230430
distlib==0.3.6
Django==5.0.1
django-async-include==0.6.10
django-extensions==3.2.3
django-fields==0.3.0
django-import-export==3.2.0
django-jsonfield==1.4.1
django-phonenumber-field==7.1.0
django-search-admin-autocomplete==0.2.1
django-simple-history==3.3.0
docx-mailmerge==0.5.0
docx2pdf==0.1.8
et-xmlfile==1.1.0
filelock==3.12.0
import-export==0.3.1
importlib-metadata==6.6.0
jsonpickle==3.0.1
lxml==4.9.3
MarkupPy==1.14
model-import-export==0.1.4
numpy
odfpy==1.4.1
openpyxl==3.1.2
pandas==2.0.3
phonenumbers==8.13.18
Pillow==10.0.0
Pillow-PIL==0.1.dev0
pipenv==2023.5.19
platformdirs==3.5.1
pycrypto==2.6.1
pycryptodome==3.18.0
python-dateutil>=2.6.0
pytz==2023.3
PyYAML==6.0.1
six==1.16.0
sqlparse==0.4.4
tablib==3.5.0
tqdm==4.65.0
typing_extensions==4.6.2
tzdata==2023.3
virtualenv==20.23.0
virtualenv-clone==0.5.7
xlrd==2.0.1
xlwt==1.3.0
zipp==3.15.0
